+++ install-stubs/Orchid/Layouts/User/UserListLayout.php
@@ -31,16 +31,25 @@ class UserListLayout extends Table
                 ->sort()
                 ->filter(TD::FILTER_TEXT)
                 ->render(function (User $user) {
-                    // Please use view('path')
-                    return "<div class='d-sm-flex flex-row flex-wrap text-center text-sm-left align-items-center'>
-                                <span class='thumb-xs avatar m-r-xs d-none d-md-inline-block'>
-                                  <img src='{$user->getAvatar()}' class='bg-light'>
-                                </span>
-                                <div class='ml-sm-3 ml-md-0 ml-xl-3 mt-2 mt-sm-0 mt-md-2 mt-xl-0'>
-                                  <p class='mb-0'>{$user->getNameTitle()}</p>
-                                  <small class='text-xs text-muted'>{$user->getSubTitle()}</small>
+                    // Please use Blade templates.
+                    // This will be a simple example: view('path', ['user' => $user])
+
+                    $avatar = e($user->getAvatar());
+                    $name = e($user->getNameTitle());
+                    $sub = e($user->getSubTitle());
+                    $route = route('platform.systems.users.edit', $user->id);
+
+                    return "<a href='{$route}}'>
+                                <div class='d-sm-flex flex-row flex-wrap text-center text-sm-left align-items-center'>
+                                    <span class='thumb-xs avatar m-r-xs d-none d-md-inline-block'>
+                                      <img src='{$avatar}' class='bg-light'>
+                                    </span>
+                                    <div class='ml-sm-3 ml-md-0 ml-xl-3 mt-2 mt-sm-0 mt-md-2 mt-xl-0'>
+                                      <p class='mb-0'>{$name}</p>
+                                      <small class='text-xs text-muted'>{$sub}</small>
+                                    </div>
                                 </div>
-                            </div>";
+                            </a>";
                 }),
 
             TD::set('email', __('Email'))