+++ resources/views/partials/search.blade.php
@@ -4,15 +4,16 @@
             <div class="input-icon">
                 <input
                        data-action="keyup->layouts--search#query blur->layouts--search#blur focus->layouts--search#focus"
-                       type="text" class="form-control input-sm  padder bg-dark text-white"
+                       data-target="layouts--search.query"
+                       type="text"
+                       class="form-control input-sm padder bg-dark text-white"
                        placeholder="{{ __('What to search...') }}"
-                       data-toggle="dropdown" aria-haspopup="true" aria-expanded="false"
                 >
                 <div class="input-icon-addon">
                     <i class="icon-magnifier"></i>
                 </div>
             </div>
-            <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow"
+            <div class="dropdown-menu dropdown-menu-right dropdown-menu-arrow bg-white w-full"
                  x-placement="start-left" id="search-result">
             </div>
         </div>