+++ src/Platform/Models/User.php
@@ -9,6 +9,7 @@ use Orchid\Screen\AsSource;
 use Orchid\Access\UserAccess;
 use Orchid\Filters\Filterable;
 use Orchid\Access\UserInterface;
+use Illuminate\Support\Collection;
 use Illuminate\Support\Facades\Hash;
 use Orchid\Support\Facades\Dashboard;
 use Illuminate\Notifications\Notifiable;
@@ -159,23 +160,22 @@ class User extends Authenticatable implements UserInterface
     }
 
     /**
-     * @return mixed
+     * @return Collection
      */
-    public function getStatusPermission()
+    public function getStatusPermission() : Collection
     {
         $permissions = $this->permissions ?? [];
 
         return Dashboard::getPermission()
             ->sort()
             ->transform(function ($group) use ($permissions) {
-                $group = collect($group)->sortBy('description')->toArray();
+                return collect($group)->sortBy('description')
+                    ->map(function ($value) use ($permissions) {
+                        $slug = $value['slug'];
+                        $value['active'] = array_key_exists($slug, $permissions) && (bool) $permissions[$slug];
 
-                foreach ($group as $key => $value) {
-                    $slug = $value['slug'];
-                    $group[$key]['active'] = array_key_exists($slug, $permissions) && (bool) $permissions[$slug];
-                }
-
-                return $group;
+                        return $value;
+                    });
             });
     }
 }