+++ src/Screen/Layouts/Base.php
@@ -57,16 +57,6 @@ abstract class Base
     protected $variables = [];
 
     /**
-     * Base constructor.
-     *
-     * @param Base[] $layouts
-     */
-    public function __construct(array $layouts = [])
-    {
-        $this->layouts = $layouts;
-    }
-
-    /**
      * @param Repository $repository
      *
      * @return mixed