+++ src/Screen/Layouts/Selection.php
@@ -19,6 +19,16 @@ abstract class Selection extends Base
     public $template = 'platform::layouts.selection';
 
     /**
+     * Base constructor.
+     *
+     * @param Base[] $layouts
+     */
+    public function __construct(array $layouts = [])
+    {
+        $this->layouts = $layouts;
+    }
+
+    /**
      * @param Repository $repository
      *
      * @return Factory|\Illuminate\View\View|mixed