+++ src/Screen/Layouts/Tabs.php
@@ -17,6 +17,16 @@ abstract class Tabs extends Base
     public $template = 'platform::layouts.tabs';
 
     /**
+     * Base constructor.
+     *
+     * @param Base[] $layouts
+     */
+    public function __construct(array $layouts = [])
+    {
+        $this->layouts = $layouts;
+    }
+
+    /**
      * @param Repository $repository
      *
      * @return mixed