+++ tests/Exemplar/App/Screens/ModalScreen.php
@@ -7,7 +7,7 @@ namespace Orchid\Tests\Exemplar\App\Screens;
 use Orchid\Screen\Link;
 use Orchid\Screen\Layout;
 use Orchid\Screen\Screen;
-use Orchid\Screen\Layouts\Modals;
+use Orchid\Screen\Layouts\Modal;
 
 class ModalScreen extends Screen
 {
@@ -66,7 +66,7 @@ class ModalScreen extends Screen
                 Layout::rows([]),
             ])
                 ->title(self::TITLE_MODAL)
-                ->size(Modals::SIZE_LG)
+                ->size(Modal::SIZE_LG)
                 ->applyButton(self::APPLY_BUTTON)
                 ->closeButton(self::CLOSE_BUTTON),
         ];